backend-fastapi/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pydantic==1.10.12
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.24.0
pypdf2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
asyncpg==0.28.0
aiohttp==3.8.5
